Extract PolygonDrawingSession into an interface and implementation
* Convert `PolygonDrawingSession` from a class to an interface.
* Create `PolygonDrawingSessionImpl` to provide the concrete implementation of the polygon drawing logic.
* Update `DrawAreaTaskViewModel` and associated tests to use the new implementation.
