+++ app/src/test/java/org/groundplatform/android/ui/datacollection/tasks/polygon/PolygonDrawingSessionTest.kt
@@ -26,7 +26,7 @@ class PolygonDrawingSessionTest {
 
   @Before
   fun setUp() {
-    session = PolygonDrawingSession()
+    session = PolygonDrawingSessionImpl()
   }
 
   @Test
